Update adding values.py

+++ adding values.py
@@ -1,13 +1,14 @@
 import pymysql
 from tkinter import *
 from tkinter import ttk
+from tkinter import messagebox
 
 try:
     db=pymysql.connect(host='localhost',user='root',password='bruh',db='grocery_store')
     print('connected')
 except Exception as e:
     print(e)
-
+#creating tkinter GUI screen
 board=Tk()
 board.geometry('900x700')
 board.title('grocery management system')
@@ -30,23 +31,50 @@ def showall():
     trv.heading("1", text ="Bar code")
     trv.heading("2", text ="Name")
     trv.heading("3", text ="Price")
-    trv.heading("4", text ="Quantity")  
+    trv.heading("4", text ="Quantity")
+    def exit2():
+     res = messagebox.askyesno('confirmation', 'Do you want to kill this window?') 
+     if res == True:
+         nb.destroy()
     try:
         cr=db.cursor()
         cr.execute('select * from store_items')
         db.commit()
         table=cr.fetchall()
     except Exception as ex:
-        print(ex)
+        messagebox.showerror('Error', ex) 
     
     for dt in table:
         trv.insert("", 'end',iid=dt[0], text=dt[0],values =(dt[0],dt[1],dt[2],dt[3]))
-    nb.mainloop()
+    b2 =Button(nb, text="EXIT", command=exit2)
+    b2.grid(row=25,column=2,padx=20,pady=20)
 
     
 def showbar():
     nb=Toplevel(board)
     v=IntVar()
+    # Using treeview widget
+    trv = ttk.Treeview(nb, selectmode ='browse')
+    trv.grid(row=1,column=1,padx=20,pady=20)
+    # number of columns
+    trv["columns"] = ("1", "2", "3","4")
+    # Defining heading
+    trv['show'] = 'headings'
+    # width of columns and alignment 
+    trv.column("1", width = 100, anchor ='c')
+    trv.column("2", width = 300, anchor ='w')
+    trv.column("3", width = 40, anchor ='c')
+    trv.column("4", width = 80, anchor ='c')
+    # Headings  
+    # respective columns
+    trv.heading("1", text ="Bar code")
+    trv.heading("2", text ="Name")
+    trv.heading("3", text ="Price")
+    trv.heading("4", text ="Quantity")
+    def exit2():
+     res = messagebox.askyesno('confirmation', 'Do you want to kill this window?') 
+     if res == True:
+         nb.destroy()
     def showit1():
         try:
             query='select * from store_items where bar_code={}'.format(v.get())
@@ -55,24 +83,32 @@ def showbar():
             db.commit()
             value=cr.fetchall()
             if value==():
-                print('does not exist')
+                messagebox.showerror('Error',"value doesn't exist. ENTER A VAID VALUE")
             else:
-                print(value)
+                for dt in value:
+                    trv.insert("", 'end',iid=dt[0], text=dt[0],values =(dt[0],dt[1],dt[2],dt[3]))
+
         except Exception as ex:
-            print(ex)
+            messagebox.showerror('Error',ex)    
 
     m = Label(nb,text = "barcode number")
     e = Entry(nb,textvariable=v)
     b=Button(nb,text='calculate',command=showit1)
-    m.pack()
-    e.pack()
-    b.pack()
-    
+    b2 =Button(nb, text="EXIT", command=exit2)
+    m.grid(row=20,column=1,padx=20,pady=20) 
+    e.grid(row=20,column=3,padx=20,pady=20)
+    b.grid(row=25,column=1,padx=20,pady=20)
+    b2.grid(row=25,column=2,padx=20,pady=20)
+    nb.mainloop()
     
 def sell():
-    nnb=Toplevel(board)
+    nb=Toplevel(board)
     v1=IntVar()
     v2=IntVar()
+    def exit2():
+     res = messagebox.askyesno('confirmation', 'Do you want to kill this window?') 
+     if res == True:
+         nb.destroy()
     def showit2():
         try:
             query1='select * from store_items where bar_code={}'.format(v2.get())
@@ -81,7 +117,7 @@ def sell():
             db.commit()
             value=cr.fetchall()
             if value==():
-                print('does not exist')
+                messagebox.showerror('Error',"value doesn't exist. ENTER A VAID VALUE")
             else:
                 print(value)
                 query2='update store_items set quantity=quantity-{} where bar_code={}'.format(v1.get(),v2.get())
@@ -89,25 +125,31 @@ def sell():
                 db.commit()
                 print('updated')
         except Exception as ex:
-            print(ex)
-    m2 = Label(nnb,text = "barcode no.")
-    e2 = Entry(nnb,textvariable=v2)
-    m1 = Label(nnb,text = "quantity sold")
-    e1 = Entry(nnb,textvariable=v1)
-    b=Button(nnb,text='calculate',command=showit2)
+            messagebox.showerror('Error', ex)
+    m2 = Label(nb,text = "barcode no.")
+    e2 = Entry(nb,textvariable=v2)
+    m1 = Label(nb,text = "quantity sold")
+    e1 = Entry(nb,textvariable=v1)
+    b=Button(nb,text='calculate',command=showit2)
+    b2 =Button(nb, text="EXIT", command=exit2)
     m2.pack()
     e2.pack()
     m1.pack()
     e1.pack()
     b.pack()
-    nnb.mainloop()
+    b2.pack()
+    nb.mainloop()
 
 
 
 def buy():
-    nnb=Toplevel(board)
+    nb=Toplevel(board)
     v1=IntVar()
     v2=IntVar()
+    def exit2():
+     res = messagebox.askyesno('confirmation', 'Do you want to kill this window?') 
+     if res == True:
+         nb.destroy()
     def showit3():
         try:
             query1='select * from store_items where bar_code={}'.format(v2.get())
@@ -116,7 +158,7 @@ def buy():
             db.commit()
             value=cr.fetchall()
             if value==():
-                print('does not exist')
+                messagebox.showerror('Error',"value doesn't exist. ENTER A VAID VALUE")
             else:
                 print(value)
                 query2='update store_items set quantity=quantity+{} where bar_code={}'.format(v1.get(),v2.get())
@@ -124,18 +166,20 @@ def buy():
                 db.commit()
                 print('updated')
         except Exception as ex:
-            print(ex)
-    m2 = Label(nnb,text = "barcode no.")
-    e2 = Entry(nnb,textvariable=v2)
-    m1 = Label(nnb,text = "quantity sold")
-    e1 = Entry(nnb,textvariable=v1)
-    b=Button(nnb,text='calculate',command=showit3)
+            messagebox.showerror('Error',ex)
+    m2 = Label(nb,text = "barcode no.")
+    e2 = Entry(nb,textvariable=v2)
+    m1 = Label(nb,text = "quantity sold")
+    e1 = Entry(nb,textvariable=v1)
+    b=Button(nb,text='calculate',command=showit3)
+    b2 =Button(nb, text="EXIT", command=exit2)
     m2.pack()
     e2.pack()
     m1.pack()
     e1.pack()
     b.pack()
-    nnb.mainloop()
+    b2.pack()
+    nb.mainloop()
 
  
 def add_new():
@@ -144,6 +188,10 @@ def add_new():
     v2=StringVar()
     v3=IntVar()
     v4=IntVar()
+    def exit2():
+     res = messagebox.askyesno('confirmation', 'Do you want to kill this window?') 
+     if res == True:
+         nb.destroy()
     def showit4():
         try:
             query1='select * from store_items where bar_code={}'.format(v1.get())
@@ -153,7 +201,7 @@ def add_new():
             value=cr.fetchall()
             if value==():
                 if v2.get()=='' or v3.get()==0 or v4.get()==0:
-                    print('fill in all the details')
+                    messagebox.showerror('Error',"some feilds are unfilled .FILL ALL THE FEILDS")
                 else:
                     query='insert into store_items values({},"{}",{},{})'.format(v1.get(),v2.get(),v3.get(),v4.get())
                     cr=db.cursor()
@@ -161,9 +209,9 @@ def add_new():
                     db.commit()
                     print('added')
             else:
-                print('already exists')
+                messagebox.showerror('Error',"value already exists")
         except Exception as ex:
-            print(ex)
+            messagebox.showerror('Error',ex)
     m1 = Label(nb,text = "barcode no.")
     e1 = Entry(nb,textvariable=v1)
     m2 = Label(nb,text = "name of the item")
@@ -173,6 +221,7 @@ def add_new():
     m4 = Label(nb,text = "quantity")
     e4 = Entry(nb,textvariable=v4)
     b=Button(nb,text='calculate',command=showit4)
+    b2 =Button(nb, text="EXIT", command=exit2)
     m1.pack()
     e1.pack()
     m2.pack()
@@ -182,6 +231,7 @@ def add_new():
     m4.pack()
     e4.pack()
     b.pack()
+    b2.pack()
     nb.mainloop()
 
 
@@ -190,6 +240,10 @@ def add_new():
 def delit():
     nb=Toplevel(board)
     v=IntVar()
+    def exit2():
+     res = messagebox.askyesno('confirmation', 'Do you want to kill this window?') 
+     if res == True:
+         nb.destroy()
     def showit5():
         try:
             query1='select * from store_items where bar_code={}'.format(v.get())
@@ -198,39 +252,46 @@ def delit():
             db.commit()
             value=cr.fetchall()
             if value==():
-                print('does not exist')
+                messagebox.showerror('Error',"value doesn't exist. ENTER A VAID VALUE")
             else:
-                query='delete from store_items where bar_code={}'.format(v.get())
-                cr=db.cursor()
-                cr.execute(query)
-                db.commit()
-                print('deleted')
+                answer = messagebox.askyesno('confirmation','Are you sure that you want to delete it?')
+                if answer:
+                    query='delete from store_items where bar_code={}'.format(v.get())
+                    cr=db.cursor()
+                    cr.execute(query)
+                    db.commit()
+                    print('deleted')
         except Exception as ex:
-            print(ex)
+            messagebox.showerror('Error', ex)
     m = Label(nb,text = "barcode number")
     e = Entry(nb,textvariable=v)
     b=Button(nb,text='calculate',command=showit5)
+    b2 =Button(nb, text="EXIT", command=exit2)
     m.pack()
     e.pack()
     b.pack()
+    b2.pack()
+    nb.mainloop()
     
-    
-
+def exit1():
+     res = messagebox.askyesno('confirmation', 'Do you want to exit this application?')
+     if res == True:
+         board.destroy()
 
 
 bb=0
-btn =Button(board, text="Press1", command=lambda:showall())
+btn =Button(board, text="show all the elements", command=showall)
 btn.pack()
-btn2 =Button(board, text="Press2", command=lambda:showbar())
+btn2 =Button(board, text="show elements using their barcode", command=showbar)
 btn2.pack()
-btn3 =Button(board, text="Press3", command=lambda:sell())
+btn3 =Button(board, text="sell items to a customer", command=sell)
 btn3.pack()
-btn4 =Button(board, text="Press4", command=lambda:buy())
+btn4 =Button(board, text="restock", command=buy)
 btn4.pack()
-btn5 =Button(board, text="Press5", command=lambda:add_new())
+btn5 =Button(board, text="add a new item to stock", command=add_new)
 btn5.pack()
-btn6 =Button(board, text="Press6", command=lambda:delit())
+btn6 =Button(board, text="delete an item from the stock", command=delit)
 btn6.pack()
-btn6 =Button(board, text="EXIT", command=board.destroy)
+btn6 =Button(board, text="EXIT", command=exit1)
 btn6.pack()
 board.mainloop()
